model build updates
