--- server/db/models/AuthenticatedUser.js
@@ -1,67 +1,0 @@
-const crypto = require('crypto')
-const Sequelize = require('sequelize')
-const db = require('../db')
-
-const AuthenticatedUser = db.define('user', {
-  email: {
-    type: Sequelize.STRING,
-    unique: true,
-    allowNull: false
-  },
-  password: {
-    type: Sequelize.STRING,
-    // Making `.password` act like a func hides it when serializing to JSON.
-    // This is a hack to get around Sequelize's lack of a "private" option.
-    get() {
-      return () => this.getDataValue('password')
-    }
-  },
-  salt: {
-    type: Sequelize.STRING,
-    // Making `.salt` act like a function hides it when serializing to JSON.
-    // This is a hack to get around Sequelize's lack of a "private" option.
-    get() {
-      return () => this.getDataValue('salt')
-    }
-  },
-  googleId: {
-    type: Sequelize.STRING
-  }
-})
-
-module.exports = AuthenticatedUser
-
-/**
- * instanceMethods
- */
-AuthenticatedUser.prototype.correctPassword = function(candidatePwd) {
-  return User.encryptPassword(candidatePwd, this.salt()) === this.password()
-}
-
-/**
- * classMethods
- */
-AuthenticatedUser.generateSalt = function() {
-  return crypto.randomBytes(16).toString('base64')
-}
-
-AuthenticatedUser.encryptPassword = function(plainText, salt) {
-  return crypto
-    .createHash('RSA-SHA256')
-    .update(plainText)
-    .update(salt)
-    .digest('hex')
-}
-
-/**
- * hooks
- */
-const setSaltAndPassword = user => {
-  if (user.changed('password')) {
-    user.salt = User.generateSalt()
-    user.password = User.encryptPassword(user.password(), user.salt())
-  }
-}
-
-AuthenticatedUser.beforeCreate(setSaltAndPassword)
-AuthenticatedUser.beforeUpdate(setSaltAndPassword)
